pyproject.toml
[project]
name = "hoopiq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiodns>=4.0.0",
    "aiohttp>=3.13.5",
    "asyncpg>=0.31.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
